requirements.txt
Django>=4.0,<5.0
djangorestframework
gunicorn
psycopg2-binary
pandas
openpyxl
